+++ NEWS.md
@@ -8,6 +8,11 @@ changes occur which require user intervention / configuration changes.  These
 are highlights since the last update, and are not meant to be an exhaustive
 listing of changes. See the git commit log for additional details.
 
+# 2021-06-xx
+
+## Backwards incompatible changes
+- Renamed `keyStorePath` to `signing.keyStorePath`.
+
 # 2021-05-04
 ## Highlights:
 - Updated vanilla flavor to RQ2A.210505